fix codecheck error

+++ services/hilogtool/main.cpp
@@ -299,7 +299,7 @@ struct HilogArgs {
     uint32_t domains[MAX_DOMAINS];
     string regex;
     string fileName;
-    uint32_t buffSize;
+    int32_t buffSize;
     uint32_t jobId;
     uint32_t fileSize;
     uint16_t levels;
@@ -540,7 +540,7 @@ static int BufferSizeSetHandler(HilogArgs& context, const char *arg)
     if (size == 0) {
         return ERR_INVALID_SIZE_STR;
     }
-    context.buffSize = size;
+    context.buffSize = static_cast<int32_t>(size);
     BufferSizeSetRqst rqst;
     context.ToBufferSizeSetRqst(rqst);
     LogIoctl ioctl(IoctlCmd::BUFFERSIZE_SET_RQST, IoctlCmd::BUFFERSIZE_SET_RSP);
